pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.120.4",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "imagekitio>=4.2.0",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.3",
]
